ml-server/pyproject.toml
[project]
name = "truegul-ml-server"
version = "0.1.0"
description = "ML Server for TrueGul - AI writing detection and feedback"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "redis>=5.0.0",
    "llama-cpp-python>=0.2.0",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
    "pre-commit>=4.0.0",
]

[tool.setuptools]
# Include the app package only
packages = ["app"]

# Ruff configuration
[tool.ruff]
line-length = 100
target-version = "py313"
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (needed for FastAPI Depends)
]

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Pyright configuration
[tool.pyright]
include = ["app"]
exclude = ["**/__pycache__"]
pythonVersion = "3.13"
typeCheckingMode = "standard"
reportMissingImports = false  # Dependencies not installed in CI check environment
reportMissingTypeStubs = false
reportUnusedImport = false  # Handled by Ruff

# Pytest configuration
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
